requirements.txt
einops==0.3.2
h5py==2.10.0
hdf5plugin==4.2.0
matplotlib==3.5.3
numba==0.53.1
numpy==1.21.6
opencv_python==4.5.2.54
opt_einsum==3.3.0
pandas
Pillow
scipy
seaborn
setuptools==57.0.0
torch==1.9.0+cu111
torch_scatter==2.0.9
torchvision==0.10.0+cu111
tqdm
typing_extensions
yacs
klampt
open3d
ouster-sdk
requests
